requirements-test.txt
# Required to run test suite
pytest==7.2.1
pytest-cov==2.12.1
requests-mock==1.8.0
mock==4.0.3
codecov==2.1.11
tox==4.4.2
